backend/requirements.txt
Django>=4.2.0,<5.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
django-cors-headers>=4.0.0
PyJWT>=2.8.0

